fastAPI/requirements.txt
anyio==3.5.0
asgiref==3.5.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.75.2
grpcio==1.44.0
grpcio-tools==1.44.0
h11==0.13.0
idna==3.3
numpy==1.23.1
Pillow==9.2.0
protobuf==3.20.1
protobuf2pydantic==2021.11.25
pycodestyle==2.8.0
pydantic==1.9.0
python-multipart==0.0.5
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
torch==1.12.0
torchvision==0.13.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6